Add return type assert support and improve performance a little.

+++ test/runtests.jl
@@ -48,13 +48,13 @@ cache = get_memos(test_static_type)[1]
     @memo a = 2x
     a * y
 end
-@memoize function test_dispatch(@memo(x::Int64), y::Float64)
+@memoize function test_dispatch(@memo(x::Int64), y::Float64)::Float64
     @memo a = 3x
     a * y
 end
 @test test_dispatch(1.0, 2.0) == test_dispatch(1.0, 2.0)
 @test test_dispatch(Int64(1), 2.0) == test_dispatch(Int64(1), 2.0)
 
-@memoize test_return(x::Int64) = x < 3 ? 1 : (test_return(x - 1) + test_return(x - 2))
+@memoize test_return(x::Int64)::Int64 = x < 3 ? 1 : (test_return(x - 1) + test_return(x - 2))
 @test test_return(Int64(5)) == 5
 @test test_return(Int64(100)) == 3736710778780434371
